Do not add duplicated entries when using COPY by passing `keys` to be checked
The `keys` arg is a list of fields. When it is passed to the fucntion, before
trying to COPY the entries to the database, it will check whether there are
entries already in the database matching the content to be added based on the
`keys` and if so, skip the duplicated ones from the COPY.

Useful to avoid problems with database uniqueness constraints.